apps/api/pyproject.toml
[project]
name = "genjishimada-api"
version = "0.1.0"
description = "Genji Shimada API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.5",
    "aiohttp>=3.12.14",
    "boto3>=1.40.25",
    "genjishimada-sdk",
    "kaleido==0.2.1",
    "litestar-asyncpg>=0.4.0",
    "litestar[standard]>=2.16.0",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.1",
    "sentry-sdk[litestar]>=2.35.1",
    "msgspec>=0.19.0",
    "rapidfuzz>=3.12.0",
    "asyncpg-stubs>=0.31.1",
    "bcrypt>=4.0.0",
    "httpx>=0.27.0",
    "sqlspec>=0.38.0",
]

[dependency-groups]
dev = [
    "psycopg[binary]>=3.3.2",
]
dev-api = [
    "httpx-sse>=0.4.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.2.0",
    "pytest-databases[postgres]>=0.14.0",
    "pytest-xdist>=3.8.0",
    "types-boto3[boto3]>=1.40.25",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: Integration tests with real database and HTTP layer",
    "domain_auth: Auth domain tests",
    "domain_autocomplete: Autocomplete domain tests",
    "domain_change_requests: Change requests domain tests",
    "domain_community: Community domain tests",
    "domain_completions: Completions domain tests",
    "domain_jobs: Jobs domain tests",
    "domain_lootbox: Lootbox domain tests",
    "domain_map_edits: Map edits domain tests",
    "domain_maps: Maps domain tests",
    "domain_newsfeed: Newsfeed domain tests",
    "domain_notifications: Notifications domain tests",
    "domain_playtests: Playtests domain tests",
    "domain_rank_card: Rank card domain tests",
    "domain_users: Users domain tests",
    "domain_utilities: Utilities domain tests",
]

[tool.ruff]
exclude = ["tests", "test_v4"]
line-length = 120
target-version = "py313"

lint.select = [
    "E",
    "F",
    "W",
    "A",
    "PL",
    "I",
    "SIM",
    "RUF",
    "ASYNC",
    "C4",
    "INP",
    "ERA",
    "SLF",
    "PIE",
    "PYI",
    "ANN",
    "N",
    "D",
]
lint.ignore = [
    "ANN002",
    "ANN003",
    "D203",
    "D100",
    "D101",
    "D104",
    "D107",
    "RUF012",
    "TC002",
    "D213",
    "TC001",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"conftest.py" = ["ALL"]
"tests/**" = ["ALL"]

[tool.basedpyright]
ignore = [
    "tests",
    "conftest.py",
]

[tool.uv.sources]
genjishimada-sdk = { workspace = true }
